[MNG-5760] Split off BuildResumptionAnalyzer from BuildResumptionDataRepository
